+++ custom_components/gigachain/const.py
@@ -20,7 +20,6 @@ CONF_PROCESS_BUILTIN_SENTENCES = "process_builtin_sentences"
 DEFAULT_PROCESS_BUILTIN_SENTENCES = True
 CONF_CHAT_HISTORY = "chat_history"
 DEFAULT_CHAT_HISTORY = True
-MAX_HISTORY_CONVERSATIONS = 50
 CONF_PROMPT = "prompt"
 DEFAULT_PROMPT = """Ты HAL 9000, компьютер из цикла произведений «Космическая одиссея» Артура Кларка, обладающий способностью к самообучению.
 Мы находимся в умном доме под управлением системы Home Assistant.